Checking in the original versions of all the HubNet activities that users can currently run.
These were taken straight from the NetLogo repository - they have not been refactored yet to have the 'go' button removed and replaced with an 'on' switch.

Once we have support for multiple users, these activities that I'm checking in here are going to be the "sample models" that all users can run.  In addition, they will be able to upload their own private activities to the server.