Adding support for basic user registration and login.  Currently, any action performed within the Activities controller requires that a user is logged in.
Using a file-based H2 database.  Some additional steps are required to get this to work on a different machine. In application.conf, the path for db.url needs to be adjusted based on where the app is located. The database has to be created first by navigating to the URL localhost:9000/@db, selecting "Generig H2 (Server)" from the drop-down list, and adjusting the JDBC URL to match what's in application.conf. Finally, the USERS table has to be created after hitting "Connect" and entering the following:

CREATE TABLE USERS(USERNAME VARCHAR(32) PRIMARY KEY, PASSWORD VARCHAR(255));